ai-service/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23

# AI/ML libraries
tensorflow==2.14.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.12.2

# Computer Vision (for face analysis)
opencv-python==4.8.1.78
mediapipe==0.10.7

# NLP (for chatbot)
transformers==4.35.2
torch==2.1.1

# API and utilities
gunicorn==21.2.0
Pillow==10.1.0
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
